tests/test_data_processing.py
"""
Test module for data processing functions.

This file should contain:
1. Tests for Macenko color normalization function
2. Tests for dataset loading and validation
3. Tests for image preprocessing pipeline
4. Tests for data augmentation transforms
5. Mock data fixtures for testing
6. Edge case testing (empty datasets, corrupted images, etc.)
7. Performance tests for large datasets
8. Integration tests with PyTorch DataLoader

Example test cases:
- test_macenko_normalize_basic()
- test_macenko_normalize_edge_cases()
- test_gram_stain_dataset_loading()
- test_dataset_transforms_application()
- test_dataset_class_balance()
"""

import pytest
import numpy as np
from PIL import Image
import torch
from unittest.mock import Mock, patch
from src.utils import normalize_macenko, MacenkoNormalize
from src.data.dataset import GramStainDataset

# Test fixtures and mock data creation
# Test cases for each function
# Integration tests
