pyproject.toml
[tool.poetry]
name = "bacterial-gan"
version = "0.1.0"
description = "Bacterial Image Augmentation using Conditional GANs"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "bacterial_gan", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
# Core ML
tensorflow = "^2.15.0"
# tensorflow-addons removed - not compatible with Python 3.12, and we don't need it
numpy = "^1.26.0"
pillow = "^10.0.0"

# MLflow & Experiment Tracking
mlflow = "^2.9.0"

# API
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
python-multipart = "^0.0.6"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"

# Background Jobs (for retraining via API)
celery = "^5.3.0"
redis = "^5.0.0"

# CLI
typer = {extras = ["all"], version = "^0.9.0"}

# Data Processing
opencv-python-headless = "^4.8.0"
scikit-image = "^0.22.0"
scikit-learn = "^1.3.0"
pandas = "^2.1.0"

# Configuration
pyyaml = "^6.0.1"
python-dotenv = "^1.0.0"

# DVC
dvc = "^3.30.0"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"

# Code Quality
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"

# Jupyter
jupyter = "^1.0.0"
ipykernel = "^6.27.0"

[tool.poetry.scripts]
bacterial-gan = "bacterial_gan.cli:app"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=html --cov-report=term-missing"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
